“Build” markup by copying files to assets
This way, the application really only concerns
itself with the assets/ directory. source/
doesn’t even have to be sent to the server on
deployment. Cleaner deployments!

This change also requires building critical
stylesheet so that pug can include it. Bonus
is that we no longer have to compile that
stylesheet at runtime, just include the already
compiled critical.css file.