Run build and test scripts from flightplan
Prefer the cleanliness of using a predeploy npm script
but it’s not very smart to wait for build and tests
before checking working directory is clean. Until we
get around to checking git working dir status from an
npm script, it’s better to let flightplan handle build
and test script commands.